Java Configuration and Profiles
