--- src/main/java/com/udit/DemoApplication.java
@@ -1,18 +1,0 @@
-package com.udit;
-
-import com.udit.Controllers.GreetingController;
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.ApplicationContext;
-
-@SpringBootApplication
-public class DemoApplication {
-
-	public static void main(String[] args) {
-
-		ApplicationContext ctx = SpringApplication.run(DemoApplication.class, args);
-
-        GreetingController greetingController = (GreetingController) ctx.getBean("greetingController");
-        greetingController.getGreeting();
-	}
-}
