+++ src/main/resources/application.properties
@@ -1,1 +1,2 @@
 server.port = 8090
+spring.profiles.active=spanish
